pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phasebridge"
version = "0.1.0"
description = "PhaseBridge — strict, lossless discrete ↔ phase ↔ discrete interchange (PIF v1)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "PhaseBridge Contributors" }]
keywords = ["phase", "signal", "interchange", "lossless", "PIF", "coherence", "kappa"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "numpy>=1.22",
]

[project.optional-dependencies]
# Install like: pip install .[msgpack,pillow,hypothesis,dev]
msgpack = ["msgpack>=1.0.4"]
pillow = ["Pillow>=9.0.0"]
hypothesis = ["hypothesis>=6.0.0"]
dev = [
  "ruff>=0.4.0",
  "mypy>=1.7.0",
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "jsonschema>=4.0.0",
]

[project.scripts]
pb-encode = "phasebridge.cli.pb_encode:main"
pb-decode = "phasebridge.cli.pb_decode:main"
pb-kappa  = "phasebridge.cli.pb_kappa:main"
pb-validate = "phasebridge.cli.pb_validate:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["phasebridge*"]

[tool.setuptools.package-data]
# If later you ship JSON/YAML schemas as package data:
# "phasebridge" = ["../schemas/*.json", "../schemas/*.yaml"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]
exclude = [
  ".git",
  ".venv",
  "build",
  "dist",
  "__pycache__",
]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
